finished steps for first scenario (new game), add scratchpad
We've created a simple interface between Player and Game. A
Player creates a Game via Player.start_game, which instantiates
a new Game object and assigns it to game attribute.

We're not thrilled about this and might amend later to something
such as:

Game.new(player)

We do like the idea of interacting with the game from the user's
perspective (Player.make_guess, etc.).