change structure between Player and Game, Player can make guess, Answer is updated
We previously used Player.start_game, but have now redesigned so that we pass
in both the Player and Answer to a Game object. This is avoid complications with
multiple players -- Player1.start_game, Player2.start_game -- would result in
different game.

We can make guesses through the Player now via Player.make_guess("hello").

This will instantiate a new Guess object, pass to Game, which will then pass
to Answer, which will update the Answer.current_answer and fill in blanks.

The Game object in this sense is a kind of handler for our other objects. We
will suffer the consequences of this when it's too late.