when player makes incorrect guess, take life and add letter to trash
Currently we're managing incorrect guesses in an array on the Answer.
In the future, we may use a Trash object instead. Game.wrong_guesses is
a proxy to Answer.wrong_guesses.

The wrong guesses array contains Guess objects as we believe it is
the responsibility of the display to decide how to display them to the
Player.