cover duplicate guesses and end-game
Had to refactor Game.check_guess so that it can handle different types
of guesses (correct, incorrent, duplicate), and so that we could handle a
message attribute.

We do not have a Display object which means we are handling messages entirely
within the Game object. Not entirely happy about this, but we'll see how it
works out.