Add Abacus for scoring system
We've added an Abacus class which will be used for our scoring system.

The class will only be concerned with returning points for a given scenario,
not for keeping a running total (responsibility of the game). Our plan is
to hopefully build this class in a way that will allow other teams to inject
the functionality into their own projects.